Fixed PointClickNav Module
Changed the startup to change its status in addition to its state to
"active" instead of "idle." There still seems to be some sort of problem
with the planning itself (on two occasions, the robot said that there
was no feasible global path when there clearly was one). It appears to
work overall though.

Changed command_server to also report a stack trace when there are errors
intializing modules. Added PointClickNav to the list of modules to load
and removed the StateLogger.