WIP Debugging plan to home
Planner is exhibiting some strange behavior - can't find solution
even when started in a goal state. It looks like for the first
two calls of planMore(100) the planner does ~100 evaluations of the
goal test but on subsequent calls it only does 2. Perhaps it
has terminated but I am somehow not checking the correct flag?